refactor: convert to ts
