+++ .gitignore
@@ -1,3 +1,6 @@
+# build / generate output
+dist
+
 # dependencies
 /node_modules
