+++ nodemon.json
@@ -1,0 +1,8 @@
+{
+  "watch": ["src"],
+  "ext": "ts",
+  "exec": "ts-node src/index.ts",
+  "events": {
+    "start": "cls || clear"
+  }
+}
